refactor!: refactor type names
