+++ Carbonate/IReactor.cs
@@ -10,9 +10,9 @@ namespace Carbonate;
 public interface IReactor
 {
     /// <summary>
-    /// Gets the ID of the event where this <see cref="Reactor"/> should respond.
+    /// Gets the ID of the subscription where this <see cref="IReactor"/> should respond.
     /// </summary>
-    Guid EventId { get; }
+    Guid Id { get; }
 
     /// <summary>
     /// Gets the name of the <see cref="IReactor"/>.
@@ -28,24 +28,19 @@ public interface IReactor
     /// Gets a value indicating whether or not the <see cref="IReactor"/> has been unsubscribed.
     /// </summary>
     /// <remarks>
-    ///     This means that the <see cref="Reactor"/> will not receive <see cref="OnReceive()"/>, <see cref="OnReceive(IMessage)"/>,
-    ///     <see cref="OnUnsubscribe"/>, or <see cref="OnError"/> invokes.
+    ///     This means that the <see cref="ReceiveReactor"/> will not receive the following notifications:
+    ///     <br/>
+    ///     <list type="bullet">
+    ///         <item><see cref="IReceiveReactor"/>.<see cref="IReceiveReactor.OnReceive()"/></item>
+    ///         <item><see cref="IReceiveReactor"/>.<see cref="IReceiveReactor.OnReceive(IMessage)"/></item>
+    ///         <item><see cref="OnUnsubscribe"/></item>
+    ///         <item><see cref="OnError"/></item>
+    ///     </list>
     /// </remarks>
     bool Unsubscribed { get; }
 
     /// <summary>
-    /// Sends the next notification without any data to the <see cref="Reactor"/> subscription.
-    /// </summary>
-    void OnReceive();
-
-    /// <summary>
-    /// Sends the next notification with the given <paramref name="message"/> to the <see cref="Reactor"/> subscription.
-    /// </summary>
-    /// <param name="message">The notification message.</param>
-    void OnReceive(IMessage message);
-
-    /// <summary>
-    /// Notifies the <see cref="Reactor"/> that the provider has finished sending push-based notifications.
+    /// Notifies the subscriber that the provider has finished sending push-based notifications and has been unsubscribed.
     /// </summary>
     /// <remarks>
     ///     Will not be invoked more than once.
@@ -53,7 +48,7 @@ public interface IReactor
     void OnUnsubscribe();
 
     /// <summary>
-    /// Notifies the <see cref="Reactor"/> that the provider has experiences an error condition.
+    /// Notifies the subscriber that the provider has experienced an error condition.
     /// </summary>
     /// <param name="error">An object that provides additional information about the error.</param>
     void OnError(Exception error);
