+++ Carbonate/IReceiveReactor.cs
@@ -1,0 +1,12 @@
+﻿// <copyright file="IReceiveReactor.cs" company="KinsonDigital">
+// Copyright (c) KinsonDigital. All rights reserved.
+// </copyright>
+
+namespace Carbonate;
+
+/// <summary>
+/// A reactor capable of standard functionality that can receive push notifications.
+/// </summary>
+public interface IReceiveReactor : IReactor, IReceiver
+{
+}
