+++ Carbonate/ReceiveReactor.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="Reactor.cs" company="KinsonDigital">
+﻿// <copyright file="ReceiveReactor.cs" company="KinsonDigital">
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
@@ -7,7 +7,7 @@ namespace Carbonate;
 /// <summary>
 /// Provides a mechanism for receiving push-based notifications.
 /// </summary>
-public sealed class Reactor : IReactor
+public sealed class ReceiveReactor : IReceiveReactor
 {
     private readonly Action? onReceive;
     private readonly Action<IMessage>? onReceiveMsg;
@@ -15,10 +15,10 @@ public sealed class Reactor : IReactor
     private readonly Action<Exception>? onError;
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="Reactor"/> class.
+    /// Initializes a new instance of the <see cref="ReceiveReactor"/> class.
     /// </summary>
-    /// <param name="eventId">The ID of the event where the <see cref="Reactor"/> responds.</param>
-    /// <param name="name">The name of the <see cref="Reactor"/>.</param>
+    /// <param name="eventId">The ID of the event where the <see cref="ReceiveReactor"/> responds.</param>
+    /// <param name="name">The name of the <see cref="ReceiveReactor"/>.</param>
     /// <param name="onReceive">Executed when a push notification occurs with no data.</param>
     /// <param name="onReceiveMsg">Executed when a push notification occurs with a message.</param>
     /// <param name="onUnsubscribe">
@@ -31,7 +31,7 @@ public sealed class Reactor : IReactor
     ///     <br/>
     ///     It is only metadata for debugging or miscellaneous purposes.
     /// </remarks>
-    public Reactor(
+    public ReceiveReactor(
         Guid eventId,
         string name = "",
         Action? onReceive = null,
@@ -39,7 +39,7 @@ public sealed class Reactor : IReactor
         Action? onUnsubscribe = null,
         Action<Exception>? onError = null)
     {
-        EventId = eventId;
+        Id = eventId;
         Name = string.IsNullOrEmpty(name) ? string.Empty : name;
         this.onReceive = onReceive;
         this.onReceiveMsg = onReceiveMsg;
@@ -48,7 +48,7 @@ public sealed class Reactor : IReactor
     }
 
     /// <inheritdoc />
-    public Guid EventId { get; }
+    public Guid Id { get; }
 
     /// <inheritdoc />
     public string Name { get; }
@@ -102,5 +102,5 @@ public sealed class Reactor : IReactor
     }
 
     /// <inheritdoc cref="object.ToString"/>
-    public override string ToString() => $"{Name}{(string.IsNullOrEmpty(Name) ? string.Empty : " - ")}{EventId}";
+    public override string ToString() => $"{Name}{(string.IsNullOrEmpty(Name) ? string.Empty : " - ")}{Id}";
 }
