+++ Testing/CarbonateTests/Helpers/Fakes/ReactableBaseFake.cs
@@ -1,0 +1,14 @@
+﻿// <copyright file="ReactableBaseFake.cs" company="KinsonDigital">
+// Copyright (c) KinsonDigital. All rights reserved.
+// </copyright>
+
+namespace CarbonateTests.Helpers.Fakes;
+
+using Carbonate;
+
+/// <summary>
+/// Used for the purpose of testing the <see cref="ReactableBase{T}"/> class.
+/// </summary>
+public class ReactableBaseFake : ReactableBase<IReactor>
+{
+}
